refactor: move cli to dedicated file
