+++ multiauth/__main__.py
@@ -1,6 +1,6 @@
 """Entry point for python cli."""
 
-from .main import cli
+from .cli import cli
 
 if __name__ == '__main__':
     cli()
